+++ MOD/Systems/MainSystems.cs
@@ -6,6 +6,7 @@ using UnityEngine;
 using Colossal.PSI.Environment;
 using System.IO;
 using Game.UI.Menu;
+using ExtendedRadio.UI;
 
 namespace ExtendedRadio.Systems;
 
@@ -43,12 +44,15 @@ public partial class MainSystem : GameSystemBase
 				canLoad = false;
 
 				List<string> ModsFolderPaths = [];
-				if(Directory.Exists(EnvPath.kLocalModsPath)) ModsFolderPaths.Add(EnvPath.kLocalModsPath);
-				if(Directory.Exists(Mod.PathToPDXMods)) ModsFolderPaths.Add(Mod.PathToPDXMods);
+				// if(Directory.Exists(EnvPath.kLocalModsPath)) ModsFolderPaths.Add(EnvPath.kLocalModsPath);
+				// if(Directory.Exists(Mod.PathToPDXMods)) ModsFolderPaths.Add(Mod.PathToPDXMods);
+				if(Directory.Exists(Mod.ModsFolderCustomRadio)) CustomRadios.RegisterCustomRadioDirectory(Mod.ModsFolderCustomRadio);
 
-				extendedRadioMono.StartCoroutine(CustomRadios.SearchForCustomRadiosFolder(ModsFolderPaths));
+				//extendedRadioMono.StartCoroutine(CustomRadios.SearchForCustomRadiosFolder(ModsFolderPaths));
 
-			}
+                Icons.LoadIconsFolder();
+
+            }
 		}
 	}
 }